split inout module. adding get_string_value to StyleBox
